requirements.txt
setuptools==59.5.0
matplotlib>=3.3.4
numpy>=1.19.5
pandas>=1.1.5
Pillow>=8.3.2
scikit-image>=0.17.2
scikit-learn>=0.24.2
tensorboard>=2.6.0
pynvml>=11.0
lmdb
ftfy
regex
tqdm~=4.64.1
pandas
timm==0.4.12
omegaconf
tabulate

--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.11.0+cu113
torchvision==0.12.0+cu113
torchaudio==0.11.0 

git+https://github.com/openai/CLIP@d50d76d#egg=clip

scipy~=1.7.3
requests~=2.28.1
